requirements.txt
qdrant-client
confluent-kafka
langgraph
langchain
langchain-community
langchain-ollama
sentence-transformers
python-dotenv
requests
pandas
